app init: print config in debug mode
For debugging it's often vital to understand
the complete rendered config that took effect.
This is a pragmatic way to pretty-print the
rendered config upon application initialization.

Only print in debug mode because in prod it
might contain sensitive values that we do not
want to end up in logs.