app config: introduce INTENDED_BASE_URL, use for api docs (#435)
* app config: introduce INTENDED_BASE_URL, use for api docs

This is motivated by #393.

Introduce new configuration variable INTENDED_BASE_URL
and use the corresponding configuration value in the
dynamically generated API documentation website.

This effectively changes the default value of the
OpenAPI spec server config from

    http://localhost:5000/

to
    http://127.0.0.1:5000/

which fixes the 'try it out' feature of the OpenAPI
docs website in the canonical dev setup ('localhost'
doesn't work for CORS reasons when the dev server
actually listens on 127.0.0.1).

The value of INTENDED_BASE_URL is going to be useful
whenever the application needs to be able to construct
an absolute URL to an API endpoint. Such as when
dynamically constructing a redirect URL for a single
sign-on flow.

* tests: adjust test_docs for INTENDED_BASE_URL

* app config: auto-heal trailing slash for INTENDED_BASE_URL

It's not always great to auto-heal bad config values
but in this specific case it's a great way to
enforce a guarantee w/o bothering users.