gha: run coveralls in container
copying the coverage data file
out of the app container after
running the test suite and
then processing it independently
resulted in various

 No source for /app/conbench/api/__init__.py

etc. Run coveralls in the app
container instead. This is a bit in
conflict with the idea that the app
container should not contain all of
those fancy test run dependencies,
ideally.