github actions: invoke test runner with docker-compose approach (#459)
* docker-compose: write coverage file to host

If in the container we run `coverage` so that
it writes to /etc/conbench-coverage-datafile in the
container via

   coverage run --data-file=/etc/conbench-coverage-datafile

then make it so that the report ends up on the host.

Can be used by CI.

* github actions: use docker-compose for running tests

So that docker-compose sets up dex.

In this initial attempt here there I try to
pragmatically resolve the conflict between
the services-provided DB vs. the docker-compose-
provided DB. Let's see if that works.

* docker-compose: up instead of run

Command should return, let's see.

* use docker-compose run --detach db

(tested locally)

* gha: run alembic via docker-compose

error was:

  connection to server at "localhost" (127.0.0.1),
  port 5432 failed: Connection refused

* log reason for create_all() invocation

* migration tests: set CREATE_ALL_TABLES to false

seems to be the mechanism to tell the app to
_not_ create DB tables upon initialization.

migration failed with

sqlalchemy.exc.ProgrammingError: (psycopg2.errors.DuplicateTable) relation "case" already exists
[SQL:
CREATE TABLE "case" (
	id VARCHAR(50) NOT NULL,
	name TEXT NOT NULL,
	tags JSONB NOT NULL,
	PRIMARY KEY (id)
)
]
(Background on this error at: https://sqlalche.me/e/14/f405)

as of what I think is a side effect of importing

conbench.db from within migrations/env.py.

* docker-compose: pass on CREATE_ALL_TABLES

* docker-compose.yaml: fix syntax

now that there's an empty item in the
enumeration I seems we need to use hyphens

* docker-compose: next attempt to set CREATE_ALL_TABLES

* gha workflow file: fix cmd syntax

* debug: shorten test run (undo before merge0

* ci: coverage: deal with file/dir confusion

error was:

    os.remove(path)
      IsADirectoryError: [Errno 21] Is a directory:
     '/etc/conbench-coverage-datafile'

* docker-compose.yml: do not set DEBUG log level

This is currently probably too noisy for
the regular CI run.

(for example, prints the Flask
app config for each test -- maybe we do
not want to generate a new app object
for each test, but this is a separate
construction site).

* gha: run coveralls in container

copying the coverage data file
out of the app container after
running the test suite and
then processing it independently
resulted in various

 No source for /app/conbench/api/__init__.py

etc. Run coveralls in the app
container instead. This is a bit in
conflict with the idea that the app
container should not contain all of
those fancy test run dependencies,
ideally.

* gha: pass on env vars into docker-compose

* gha: disable migrations test for quicker debugging

* gha: coverage: use coverallsapp/github-action

* gha: do not run db (for debug)

* gha: install deps later

* gha: LCOV conversion in its own step

* gha: skip lint for quicker debug

* gha: generate more coverage data, explicit cov data file path

* gha: lint again

Install lint dependencies quickly.
We may want to pin those versions,
maybe not. The lint result should
be here w/o having to install
bigger dependencies.

* gha: dep install closer to where needed

* gha: less tests, for debug

* gha: fix yaml syntax

* gha: remove debug shortcuts

* gha: lint first, then test migrations