config: introduce CONBENCH_OIDC_ISSUER_URL  (#457)
* config: introduce CONBENCH_OIDC_ISSUER_URL

An OpenID Connect (OIDC) issuer URL precisely identifies
an OIDC identity provider. Make this configurable, for
testing and more flexible deployments.

This is the same as PR #454 which got reverted.

* oidc: dynamically generate callback URL

The previous state implicitly _required_
INTENDED_BASE_URL to be set to a meaningful
value. That was not true for old deployments.

Use the previous method again based on
Flask's url_for() which dynamically
constructs the absolute URL using data
from the incoming HTTP request.

* cleanup

* abs_oidc_callback_url: also generate scheme dynamically

based on what the client used (or at least, based
on what we think the client used).