app/_endpoint.py: rework public_data_off()
This is just a code clarity change that
is not supposed to change behavior.

This basically just resembles my findings
while trying to understand the current
the current access control model, and lays
out an idea for what to change once the
redirection to the login page must contain
the 'intended target URL'.