buildkite: disable conbench-deploy trigger
Effectively, this disables auto-deploy
to conbench.ursa.dev upon a merge-to-main.

In more specific terms, this disables
triggering the pipeline

  apache-arrow/conbench-deploy

at the end of pipeline

  apache-arrow/conbench-test

(when the branch name is `main`).

What's not visible in this commit is
that we add a new build trigger for
another buildkite pipeline that will
deploy after merge-to-main into a
more appropriate staging environment.