config: make INTENDED_BASE_URL required (#494)
* tests: set unique APPLICATION_NAME

This is towards making usage of Conbench's config
interface in the test suite.

* config: towards singleton object

Make `Config` be a singleton object, i.e.
an instantiated class instead of just a
class object.

This allows for using the class instance
initialization phase to do business
logic. Do this kind of business logic
for he OIDC_ISSUER_URL.

* config: add comment for CREATE_ALL_TABLES

* config: remove TestConfig class

Given the specific properties set
I got the impression it's not actually
needed. Seems to be the case, tests
pass.

* config: move intended base URL validation logic

This does not change behavior

* config: make INTENDED_BASE_URL required

* isort

* update application name in _expected_docs.py

* docker-compose: set INTENDED_BASE_URL for migration svc