oidc: introduce parse_state_from_authorization_response()
The business logic in the callback endpoint
became spaghetti code. Factor this part of the
work out in its own function.

The two input parameters are now documented
nicely, also via type annotations.

At the heart this patch includes a change from

    parse_request_uri_response(f.request.url)

to

    parse_request_uri_response(cur_request_url_abs)

which addresses an oautlib insecure_transport error.