docker-compose: make app listen on host, dynamic port
This is useful for running the app while exposing
it's main listen port on the host network. This
removes the implicit assumption that port 5000
can always be listened on.

The syntax used here (127.0.0.1::5000) has evolved
quite a bit between docker compose schema versions
and so I am also hard-coding 3.9 here, the documentation
of which I have read to get to this behavior.