docker-compose: add health check for app
This is useful for

    docker compose up --wait --detach

Note that this only works for `docker compose`
and not `docker-compose`.

Also note that this health check is executed
periodically, not just once. The period is
defined by `interval`. The first execution
takes `interval` amount of time. This is
a bid sad, as also discussed here:

https://forums.docker.com/t/healthcheck-different-interval-for-startup-check-vs-recurring-checks-maybe-make-a-separate-startupcheck/31318

Also reduce the interval for postgres so
that in the common case the containers appear
to be healthy rather quickly after startup.