makefile: cmd to recreate expected api docs (#508)
* docker-compose: make app listen on host, dynamic port

This is useful for running the app while exposing
it's main listen port on the host network. This
removes the implicit assumption that port 5000
can always be listened on.

The syntax used here (127.0.0.1::5000) has evolved
quite a bit between docker compose schema versions
and so I am also hard-coding 3.9 here, the documentation
of which I have read to get to this behavior.

* docker-compose: add health check for app

This is useful for

    docker compose up --wait --detach

Note that this only works for `docker compose`
and not `docker-compose`.

Also note that this health check is executed
periodically, not just once. The period is
defined by `interval`. The first execution
takes `interval` amount of time. This is
a bid sad, as also discussed here:

https://forums.docker.com/t/healthcheck-different-interval-for-startup-check-vs-recurring-checks-maybe-make-a-separate-startupcheck/31318

Also reduce the interval for postgres so
that in the common case the containers appear
to be healthy rather quickly after startup.

* docker-compose: explicitly note postgres port

This makes it so that _within_ the network
individual containers (such as `app`) can reach
the postgres container under said port.

Note that this does not expose the postgres
container port to the host.

* makefile: introduce rebuild-expected-api-docs and run-app

This is an experiment. The command

    make rebuild-expected-api-docs

might become useful in the future.

* makefile: display diff