devprod: mount local checkout into container
This introduces a new makefile target which
invokes the docker-compose machinery in a
special way, using an extension config
on top of the normal config.

It also tweaks the gunicorn settings,
see commentary. For dev I reduced the
number of workers from 5 to 2. Not just 1,
so that during local dev one can also see
some concurrency issues that would otherwise
happen in prod. Found one right away:
`db.create_all()` was not guarded against
concurrent execution. This did not happen
so far because gunicorn was started with
--preload. I removed --preload to make
the reload-upon-file-change more
complete (init code changes would
otherwise _not_ lead to auto-reload).

Why not always mount local checkout into
/app when doing `make app-run`?

Maybe that would have been easier and
also possible. But then, `make app-run`
is not really testing the contents of the
container image defined by Dockerfile
(which is the image that's being used in prod).

I feel like mounting the local checkout
into the container image is special enough
of a use case to get its own makefile target.

`make run-app` is mainly for users who 'just'
want to try out the app locally.