Accept `"pr_number": ""`
Accept `"pr_number": ""` (in addition to `"pr_number": null` and not populating it) when posting runs and results (`run.github.pr_number` and `benchmark.github.pr_number`), as this metadata is very likely to come from env vars and the intent here is unambiguous. I don't think passing `""` is desired behavior, but it's a worse reason for failing.