k8s: build custom version of kube-prometheus
For actually injecting the custom dashboard.
Did more investigation. All other dashboards
are acutally injected via volume mounts, and
these volumes are based on configmaps.

We of course could retain this method for
injection of the custom dashboard(s), too.
But it feels like we should follow the
recommended way then. This adds the JSONNET
complexity to the build chain, but it's
sane / robust tooling.