UI: landing page: remove jumbotron
With the link to the project in the
footer I don't think we need this huge
branding piece anymore.

this also cleans up the spacing/padding a bit
via p-2 -> p-0.

brings back the entry count selection menu to
the top, it fits again I think, and is simply
more accessible.