UI: multi-repo embracement; recent runs table per-repo, skip API reserialization roundtrip (#891)
* commit.py: add clearer properties

* commit.py: comments/warning about repo url

* landing page: re-structure query

- do not go through API layer anymore (save one
  re-serialization roundtrip). This is for #890.

- this is not a functional commit

- use new query technique, has great typing result

* landing page: group runs by repo

* make db-populate: inject multi-repo data

* index.html: re-build 'recent runs' table, one for each repo

* index.html: DT init: once for each table

* index.html: olumn_search_implementation() once per table

* index.html: disable legacy feature

* run.html: JS: adjust to new signature

Inject element (as returned by jquery method based
on CSS selector), not just the CSS selector

* app.html: adjust JS func signatures

arg is now element, not just selector string

* app.html: disable legacy feature

We will revisit that with something nicer
This is not tested, not documented.
And right now it gets in the way a bit.

* lint

* lint

* ui: landing page: show repo name (not URL)

* index.html: tweak commentary, remove 'sortable' code

Which columns should be sortable? well, the columns
changed order. Remove this for now, we can re-add this
in another path if it's wanted.

* lint

* index.py: sort repos alphabetically

* html templates: remove enable_search_query_string() for now

* ci: make db-populate: fix outout interpretation

it started to emit more than one ID. that's fine.
just pick the first one for screenshotting.