sed expression: use special delimiter
when using s/foo/bar/g I am used to the idea that we
can use a different delimiter like in s|foo|far|g.

But with the "address range" or "pattern range"
or "range match" technique (which is _not_ the sed
substitute command), this doesn't work naively.

Here, the slash is still the regex separator /regex/

> Note that this does not work when the sed expression begins with the delimiter,

(https://backreference.org/2010/02/20/using-different-delimiters-in-sed/index.html#comment-25209)

and then smartypants gladly joins to say

> It just has to be escaped, as explained in the article and in the sed docs.