webapp: add BMRT cache
Implement a simple job which populates and provides
a cache for the N most recent benchmark results (the
tail end of what was inserted).

The update occurs periodically.

The cache will be used by a new UI section that
is all about "listing conceptual benchmarks"
(their results, grouped by cases, hardware, context).
These pages cannot be built in a snappy way when
basing them on a meaningful amount of data _and_
serving everything freshly from the DB.

The query takes ~10 seconds on my machine against
(a local copy of) the Arrow DB.

This cache uses (quite a bit of) RAM. Start
conservatively. On my laptop with 32 GB RAM
I could put 10^6 items into that cache easily,
but we run this as of today with memory constraints.

Of course we may want to use __slots__ to reduce
the footprint.

Terminology: benchmark result tail-end (BMRT)