bmrt cache job: invoke in gunicorn hook
Initiate the BMRT cache populuation logic
in a gunicorn worker process hook (_after_
having initialized the WSGI app).

That means that simply importing the application
object or initializing the Flask app itself
does not have the side effect of running this
job.

This makes it so that running the test suite
with the pytest test runner does not initiate
the job.

With that, we can remove the

    if Config.TESTING:

bandaid again.

The result is that the job starts when
running

   make run-app-dev

locally (desired outcome). I think we now could
also remove the CREATE_ALL_TABLES-based check,
but it's OK to stay for a little longer.

This patch also invokes the original signal handlr for
SIGINT which on CPython is to raise KeyboardInterrupt.