Tighten API spec with respect to GitHub commit information, adjust tests (#1134)
* API schema: github commit: update repo&hash spec

* db schema: nullable many-to-one for run-to-commit

Allow for a Run entity to _not_ point to a
Commit, which is the semantically
cleaner way to say that it is not
associated with a commit (as compared to
using a special "no commit" Commit object.

* test: do not send commit info, instead of empty data

This is the 'correct' way to not send 'github commit'
info. To not include the object.

* add imports

* remove create_no_context() method

that is now not needed anymore,
this special commit state can
go away

* cleaner var names

the staticmethod diff is a noop

* tests: adjust to rename

* tests: adjust assert_non_commit

* commit.py: add type TypeCommitInfoGitHub

Look at those guarantees. So important
for writing simpler code.

* github commit object: implement more validation

so that we can make use of some constraints
in downstream code.

* Run: add method associated_commit_repo_url

use it for metrics

* simplify

* run.py: further simplify

* run.py: rollback before fetch

* run.py: further adjust

* commit.py: rename, use types

* rename schema

* fix typo

* tests: result DB insertion: add comment, do schema-based augment

* fix string

* github commit schema: change_empty_string_to_none()

* github commit schema: return TypeCommitInfoGitHub

* adjust to schema rename

* run.py: use new schema.load() return type

* benchmark_result.py: use new type

* commit.py: add type ignore

* tests: expect Bad Request responses

So, these test changes show that
the input parameter space was
nicely narrowed down.

* add more docs

* tests: adjust to new interface

* adjust tests

* rebuild docs

* aaaaaaah: adjust test.

this was tough. tough. tougher than it
should have been.

* lint

* lint: ignore

* explicitly handle cases where no commit info is set for a run/result

All these code sections were pointed out by mypy.
These are the fruits of this effort.

* two more type check fixes

* adjust tests to one result less, position 13 (1-index)

* rename var

* add clarity

* defuse test

* undo bad type annotation: pr_number is int

* log more detail for unexpected response

* maybe this is cleaner

* one None result went away

* restore previous test

this was an accidental WIP commit reflecting my
debugging frustration :)

* add migration for nullable foreign key

* result submission: repo URL: accept git@github.com

Accept this notation for now,
see PR discussion.

* remove test

* remove stray log message