guarantee BenchmarkResult.mean again, add clarity in many ways, introduce 'single value summary' to feed plot and history api (#1172)
* units.py and history.py: more mean-related safeguards

* rework checks, use math.isnan()

Important:

>>> math.isnan('rofl')
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
TypeError: must be real number, not str

* fix

* fix

* improve comments

* plots/results: pass BenchmarkResult object instead of dict

This is part of the ongoing effort to have less
indirection through the API layer, and access
to the convenient properties of the proper
BenchmarkResult object's properties in _plots.py

* tests: compare z-score values with leeway

Not 100 % sure where tiny deviations have their
origin, but for the z-score we do not need precision
across 10 orders of magnitude.

* rename: _to_float() -> _to_float_or_none()

* plots: require same unit for all results in history

* test: cover mean is always set

This is mainly for catching the regression
tracked by #1169

* BenchmarkResult create: calculate mean for non-errored results

* change comment for BenchmarkResult.mean

* update comment for BenchmarkResult.unit

* BenchmarkResult: update type annotation for data/times

with the SQLAlchemy Numeric option this is documented
to return decimal.Decimal, not float. That is where
a lot of confusion came from. Woof.

* BenchmarkResult: add svs()/is_failed() methods

* cleanup

* remove dead code

* add string representation to historysample

* benchmark result data: str/float confusion. woof.

* import Union

* fix assert_equal_leeway()

* historysample: add svs / svs_type

* add svs key where needed

* history.py: add function execute_history_query_get_dataframe(statement)

* history.py: use  execute_history_query_get_dataframe()

* missing import

* history.py: use svs instead of mean for analysis

svs currently _is_ mean, but this prepares for
a dynamic metric to do the analysis on.

* fix a type error

* re-activate magic test

* history plot: use svs

This does not yet dynamically read out
svs_type, but that will be an easy thing
to do next.

* cleaner comment

* adjust expected docs