BMRT: add dataclass w/ __slots__ for mem reduction
Coarse-grained experimentation shows
that this may reduce memory usage by
two orders of magnitude. Previously, the
BenchmarkResult objects retained in the
BMRT cache were (known to be) huge because
of a large amount of overhead (some of
that as of SQLAlchemy integration).

Expunging / making transient (removing from
SQLAlchemy session) already helped a lot.
Deleting individual properties helped more.
But going to a __slots__-using dataclass
helped much more.

Keeping 60.000 objects in memory resulted in
~three GB heap size in a certain deployment,
which is, well, a little ridiculous. This
here takes the next step.