run view: never show history plot, outliers
This hasn't been used really. It is also
hard to maintain (goals unclear).

Can certainly re-introduce this when desired
and when better defined.

For now, I look at this view as "display
information about a single run".

With links to more interesting analyses.
No plots on this page.
Yes: show benchmark results in this run,
and metadata about run.