value/unit/fmt cleanup (fix #1325),  misc (#1326)
* remove test_units, move _simple_source()

The unit conversion is error-prone.
Let's remove complexity until later.

The _simple_source has a misleading name.
Its code is hard to read. Move to
local context.

* plots: fmt funcs based on numstr

* Remove units.py.

That's a relieving move. fmt_unit() has created pain.

* _util.py: use numstr, not formatter_for_unit()

* _plots.py: remove 'formatted' arg

To date I did not understand what this was for.

Removing this allows for slowly adding sanity
to _source().

* remove test_plots

This tests features that are not used anymore:

- "should format"? this is not asked by code anymore,
  but simply instructed (give me the "formatted value"
  in string with certain precision and unit, as opposed to float)
- number truncation (we use a new method based on sigfig count)
- unit conversion: for now deliberately removed