UI: more control over DT column width
Datatables column width rendering is complex.
It respects some CSS, but attempts to
re-calculate during adjust(), trying to satisfy
a number of constraints of which CSS is
just one type of constraint.

With `word-break: break-all;` applied to a set
of columns there is more freedom for the
algorithm to satisfy more constraints.

related:
https://datatables.net/reference/option/columns.width