UI: tables: dynamically switch to scientific notation, better spacing (#1396)
* numstr: add numstr_dyn and numstr_exp

* UI: more control over DT column width

Datatables column width rendering is complex.
It respects some CSS, but attempts to
re-calculate during adjust(), trying to satisfy
a number of constraints of which CSS is
just one type of constraint.

With `word-break: break-all;` applied to a set
of columns there is more freedom for the
algorithm to satisfy more constraints.

related:
https://datatables.net/reference/option/columns.width

* UI: use numstr_dyn in two places

It's not yet used consistently everywhere
e.g. not in the comparison table for
resolving #1380 -- that needs more work.
But it's a good start towards addressing
#1380.

I think it's better to use scientific notation
when applicable instead of trying to dynamically
switch units.