grafana dash: add 'cluster' and 'velox' support (#1408)
"Ad-hoc variables" are a Grafana concept that
seemingly allows for addding a label filter
to _all_ queries in a dashboard (applies to
all panels). This is the second time that I
failed making nice usage of that concept.

I fell back to using a custom variable with
hard-coded values, and then manually adjusting
the query expression of each query.
Well well.