ui: res-res-cmp: show hardware checksum
This is a naughty change. Changes UI view
at the cost of changing behavior of a JSON
API endpoint. Adding a field hopefully does not
break anything.

Why does this work? The new key is picked
up func all_keys() in app/compare.py (which is
called by macro attribute_fields() in the
compare-entity jinja template).


/