UI: result-result comparison: show hardware checksum (#1415)
* ui: res-res-cmp: show hardware checksum

This is a naughty change. Changes UI view
at the cost of changing behavior of a JSON
API endpoint. Adding a field hopefully does not
break anything.

Why does this work? The new key is picked
up func all_keys() in app/compare.py (which is
called by macro attribute_fields() in the
compare-entity jinja template).


/

* start adjusting tests

this will be madness

* ui: display hardware checksum without changing API

While working myself to update tests I realized
that I need to short-cut this with a different
approach.