BMRT cache population, c-bench UI: start adding test infra and tests (#1427)
* job.py: keep track of thread objects

And allow for joining them explicitly.

* metrics.py: expose thread object

* wip: c-bench tests

* wip: test work

* add comment: issue for batch_id required

* job.py: snappy start/stop, event propagation

Add a pragmatic interface for precise control from
testing. The special and rather limiting aspect here
is that the test runner is in the main thread, in
the same thread as the HTTP handlers. This challenge
is tracked in #1431.

* add comment

* tests: cover BMRT cache population and /c-benchmarks behavior

super basic start, but this is already pretty valuable

* lint

* job.py: maintain simple return value

threading.Thread, pull other branch
outside