ci: kube-prom: jb install from local dir (respect pyrra version pin) (#1494)
ci: kube-prom: jb install from local dir

This changes from

    jb install github.com/orig/repo/dir/@commit

to

    jb install

after clone & checking out the desired commit. This changes

This changes dependency resolution in a subtle way so that after all
the pyrra version pin as specified in jsonnetfile.lock.json at the
kube-prometheus repo root is respected.

See
https://github.com/prometheus-operator/kube-prometheus/issues/2224
for more detail.

This is to resolve #1493.